+++ cumulusci/salesforce_api/metadata.py
@@ -170,7 +170,7 @@ class BaseMetadataApiCall(object):
             if refresh:
                 self.org_config.refresh_oauth_token()
                 return self._call_mdapi(headers, envelope, refresh=False)
-        # Log the error on the PackageInstallation
+        # Log the error
         message = '{}: {}'.format(faultcode, faultstring)
         self._set_status('Failed', message)
         raise MetadataApiError(message, response)
@@ -204,10 +204,10 @@ class BaseMetadataApiCall(object):
         else:
             # If no done element was in the xml, fail logging the entire SOAP
             # envelope as the log
-            self._set_status('Failed', response.content)
+            self._set_status('Failed', response.content, response=response)
         return response
 
-    def _set_status(self, status, log=None, level=None):
+    def _set_status(self, status, log=None, level=None, response=None):
         if not level:
             level = 'info'
             if status == 'Failed':
@@ -218,6 +218,9 @@ class BaseMetadataApiCall(object):
             logger('[{}]: {}'.format(status, log))
         else:
             logger('[{}]'.format(status))
+
+        if level == 'error':
+            raise MetadataApiError(log, response)
 
 
 class ApiRetrieveUnpackaged(BaseMetadataApiCall):